moving all condensers to separate files
