+++ openhands/memory/condenser/condenser.py
@@ -4,23 +4,11 @@ from abc import ABC, abstractmethod
 from contextlib import contextmanager
 from typing import Any
 
-from litellm import supports_response_schema
-from pydantic import BaseModel
 from typing_extensions import override
 
 from openhands.controller.state.state import State
-from openhands.core.config.condenser_config import (
-    AmortizedForgettingCondenserConfig,
-    CondenserConfig,
-    LLMAttentionCondenserConfig,
-    LLMSummarizingCondenserConfig,
-    ObservationMaskingCondenserConfig,
-    RecentEventsCondenserConfig,
-)
-from openhands.core.logger import openhands_logger as logger
+from openhands.core.config.condenser_config import CondenserConfig
 from openhands.events.event import Event
-from openhands.events.observation import AgentCondensationObservation, Observation
-from openhands.llm.llm import LLM
 
 CONDENSER_METADATA_KEY = 'condenser_meta'
 """Key identifying where metadata is stored in a `State` object's `extra_data` field."""
@@ -182,263 +170,3 @@ class RollingCondenser(Condenser, ABC):
         self._last_history_length = len(state.history)
 
         return results
-
-
-class ObservationMaskingCondenser(Condenser):
-    """A condenser that masks the values of observations outside of a recent attention window."""
-
-    def __init__(self, attention_window: int = 5):
-        self.attention_window = attention_window
-
-        super().__init__()
-
-    def condense(self, events: list[Event]) -> list[Event]:
-        """Replace the content of observations outside of the attention window with a placeholder."""
-        results: list[Event] = []
-        for i, event in enumerate(events):
-            if (
-                isinstance(event, Observation)
-                and i < len(events) - self.attention_window
-            ):
-                results.append(AgentCondensationObservation('<MASKED>'))
-            else:
-                results.append(event)
-
-        return results
-
-    @classmethod
-    def from_config(
-        cls, config: ObservationMaskingCondenserConfig
-    ) -> ObservationMaskingCondenser:
-        return ObservationMaskingCondenser(**config.model_dump(exclude=['type']))
-
-
-ObservationMaskingCondenser.register_config(ObservationMaskingCondenserConfig)
-
-
-class RecentEventsCondenser(Condenser):
-    """A condenser that only keeps a certain number of the most recent events."""
-
-    def __init__(self, keep_first: int = 0, max_events: int = 10):
-        self.keep_first = keep_first
-        self.max_events = max_events
-
-        super().__init__()
-
-    def condense(self, events: list[Event]) -> list[Event]:
-        """Keep only the most recent events (up to `max_events`)."""
-        head = events[: self.keep_first]
-        tail_length = max(0, self.max_events - len(head))
-        tail = events[-tail_length:]
-        return head + tail
-
-    @classmethod
-    def from_config(cls, config: RecentEventsCondenserConfig) -> RecentEventsCondenser:
-        return RecentEventsCondenser(**config.model_dump(exclude=['type']))
-
-
-RecentEventsCondenser.register_config(RecentEventsCondenserConfig)
-
-
-class LLMSummarizingCondenser(Condenser):
-    """A condenser that relies on a language model to summarize the event sequence as a single event."""
-
-    def __init__(self, llm: LLM):
-        self.llm = llm
-
-        super().__init__()
-
-    def condense(self, events: list[Event]) -> list[Event]:
-        """Applies an LLM to summarize the list of events.
-
-        Raises:
-            Exception: If the LLM is unable to summarize the event sequence.
-        """
-        try:
-            # Convert events to a format suitable for summarization
-            events_text = '\n'.join(f'{e.timestamp}: {e.message}' for e in events)
-            summarize_prompt = f'Please summarize these events:\n{events_text}'
-
-            resp = self.llm.completion(
-                messages=[{'content': summarize_prompt, 'role': 'user'}]
-            )
-            summary_response = resp.choices[0].message.content
-
-            # Create a new summary event with the condensed content
-            summary_event = AgentCondensationObservation(summary_response)
-
-            # Add metrics to state
-            self.add_metadata('response', resp.model_dump())
-            self.add_metadata('metrics', self.llm.metrics.get())
-
-            return [summary_event]
-
-        except Exception as e:
-            logger.error(f'Error condensing events: {str(e)}')
-            raise e
-
-    @classmethod
-    def from_config(
-        cls, config: LLMSummarizingCondenserConfig
-    ) -> LLMSummarizingCondenser:
-        return LLMSummarizingCondenser(llm=LLM(config=config.llm_config))
-
-
-LLMSummarizingCondenser.register_config(LLMSummarizingCondenserConfig)
-
-
-class AmortizedForgettingCondenser(RollingCondenser):
-    """A condenser that maintains a condensed history and forgets old events when it grows too large."""
-
-    def __init__(self, max_size: int = 100, keep_first: int = 0):
-        """Initialize the condenser.
-
-        Args:
-            max_size: Maximum size of history before forgetting.
-            keep_first: Number of initial events to always keep.
-
-        Raises:
-            ValueError: If keep_first is greater than max_size, keep_first is negative, or max_size is non-positive.
-        """
-        if keep_first >= max_size // 2:
-            raise ValueError(
-                f'keep_first ({keep_first}) must be less than half of max_size ({max_size})'
-            )
-        if keep_first < 0:
-            raise ValueError(f'keep_first ({keep_first}) cannot be negative')
-        if max_size < 1:
-            raise ValueError(f'max_size ({keep_first}) cannot be non-positive')
-
-        self.max_size = max_size
-        self.keep_first = keep_first
-
-        super().__init__()
-
-    def condense(self, events: list[Event]) -> list[Event]:
-        """Apply the amortized forgetting strategy to the given list of events."""
-        if len(events) <= self.max_size:
-            return events
-
-        target_size = self.max_size // 2
-        head = events[: self.keep_first]
-
-        events_from_tail = target_size - len(head)
-        tail = events[-events_from_tail:]
-
-        return head + tail
-
-    @classmethod
-    def from_config(
-        cls, config: AmortizedForgettingCondenserConfig
-    ) -> AmortizedForgettingCondenser:
-        return AmortizedForgettingCondenser(**config.model_dump(exclude=['type']))
-
-
-AmortizedForgettingCondenser.register_config(AmortizedForgettingCondenserConfig)
-
-
-class ImportantEventSelection(BaseModel):
-    """Utility class for the `LLMAttentionCondenser` that forces the LLM to return a list of integers."""
-
-    ids: list[int]
-
-
-class LLMAttentionCondenser(RollingCondenser):
-    """Rolling condenser strategy that uses an LLM to select the most important events when condensing the history."""
-
-    def __init__(self, llm: LLM, max_size: int = 100, keep_first: int = 0):
-        if keep_first >= max_size // 2:
-            raise ValueError(
-                f'keep_first ({keep_first}) must be less than half of max_size ({max_size})'
-            )
-        if keep_first < 0:
-            raise ValueError(f'keep_first ({keep_first}) cannot be negative')
-        if max_size < 1:
-            raise ValueError(f'max_size ({keep_first}) cannot be non-positive')
-
-        self.max_size = max_size
-        self.keep_first = keep_first
-        self.llm = llm
-
-        # This condenser relies on the `response_schema` feature, which is not supported by all LLMs
-        if not supports_response_schema(
-            model=self.llm.config.model,
-            custom_llm_provider=self.llm.config.custom_llm_provider,
-        ):
-            raise ValueError(
-                "The LLM model must support the 'response_schema' parameter to use the LLMAttentionCondenser."
-            )
-
-        super().__init__()
-
-    def condense(self, events: list[Event]) -> list[Event]:
-        """If the history is too long, use an LLM to select the most important events."""
-        if len(events) <= self.max_size:
-            return events
-
-        target_size = self.max_size // 2
-        head = events[: self.keep_first]
-
-        events_from_tail = target_size - len(head)
-
-        message: str = """You will be given a list of actions, observations, and thoughts from a coding agent.
-        Each item in the list has an identifier. Please sort the identifiers in order of how important the
-        contents of the item are for the next step of the coding agent's task, from most important to least
-        important."""
-
-        response = self.llm.completion(
-            messages=[
-                {'content': message, 'role': 'user'},
-                *[
-                    {
-                        'content': f'<ID>{e.id}</ID>\n<CONTENT>{e.message}</CONTENT>',
-                        'role': 'user',
-                    }
-                    for e in events
-                ],
-            ],
-            response_format={
-                'type': 'json_schema',
-                'json_schema': {
-                    'name': 'ImportantEventSelection',
-                    'schema': ImportantEventSelection.model_json_schema(),
-                },
-            },
-        )
-
-        response_ids = ImportantEventSelection.model_validate_json(
-            response.choices[0].message.content
-        ).ids
-
-        self.add_metadata('all_event_ids', [event.id for event in events])
-        self.add_metadata('response_ids', response_ids)
-        self.add_metadata('metrics', self.llm.metrics.get())
-
-        # Filter out any IDs from the head and trim the results down
-        head_ids = [event.id for event in head]
-        response_ids = [
-            response_id for response_id in response_ids if response_id not in head_ids
-        ][:events_from_tail]
-
-        # If the response IDs aren't _long_ enough, iterate backwards through the events and add any unfound IDs to the list.
-        for event in reversed(events):
-            if len(response_ids) >= events_from_tail:
-                break
-            if event.id not in response_ids:
-                response_ids.append(event.id)
-
-        # Grab the events associated with the response IDs
-        tail = [event for event in events if event.id in response_ids]
-
-        return head + tail
-
-    @classmethod
-    def from_config(cls, config: LLMAttentionCondenserConfig) -> LLMAttentionCondenser:
-        return LLMAttentionCondenser(
-            llm=LLM(config=config.llm_config),
-            max_size=config.max_size,
-            keep_first=config.keep_first,
-        )
-
-
-LLMAttentionCondenser.register_config(LLMAttentionCondenserConfig)
